Fix various other linting issues after upgrade of clang-tidy

+++ src/raytrace/test/resource_tests.cpp
@@ -13,7 +13,7 @@ TEST(GeometryTest, ObjLoad) {
 
     Mesh* mesh = resources.getMesh(id);
 
-    int num_verts = mesh->attrib.vertices.size() / 3;
+    uint64_t num_verts = mesh->attrib.vertices.size() / 3;
     EXPECT_EQ(num_verts, 7958) << "Loaded incorrect number of vertices";
 
     // assumes one shape
@@ -31,7 +31,7 @@ TEST(GeometryTest, MatLoad) {
 
     optix::Material mat = resources.getMaterial(mat_id);
 
-    int num_vars = mat->getVariableCount();
+    uint64_t num_vars = mat->getVariableCount();
     EXPECT_EQ(num_vars, 1) << "Loaded incorrect number of variables";
     renderer->getContext()->destroy();
     delete renderer;
@@ -46,10 +46,10 @@ TEST(GeometryTest, GeometryInstance) {
 
     optix::GeometryInstance instance = resources.createGeometryGroup(mesh_id, mat_id)->getChild(0);
 
-    int num_vars = instance->getVariableCount();
+    uint64_t num_vars = instance->getVariableCount();
     EXPECT_EQ(num_vars, 4) << "Did not load buffer and material variables";
 
-    int num_mats = instance->getMaterialCount();
+    uint64_t num_mats = instance->getMaterialCount();
     EXPECT_EQ(num_mats, 1) << "Did not load correct number of materials";
 
     optix::Material mat = instance->getMaterial(0u);
